Refactor remaining exit to removeLiquidity
