+++ test/weightedExit.integration.test.ts
@@ -29,11 +29,11 @@ import {
 } from '../src';
 import { forkSetup } from './lib/utils/helper';
 import {
-    assertProportionalExit,
-    assertSingleTokenExit,
-    assertUnbalancedExit,
-    doExit,
-} from './lib/utils/exitHelper';
+    assertRemoveLiquidityProportional,
+    assertRemoveLiquiditySingleToken,
+    assertRemoveLiquidityUnbalanced,
+    doRemoveLiquidity,
+} from './lib/utils/removeLiquidityHelper';
 import { RemoveLiquidityTxInput } from './lib/utils/types';
 import { ANVIL_NETWORKS, startFork } from './anvil/anvil-global-setup';
 
@@ -42,7 +42,7 @@ const { rpcUrl } = await startFork(ANVIL_NETWORKS.MAINNET);
 const poolId =
     '0x5c6ee304399dbdb9c8ef030ab642b10820db8f56000200000000000000000014'; // 80BAL-20WETH
 
-describe('weighted exit test', () => {
+describe('weighted remove liquidity test', () => {
     let txInput: RemoveLiquidityTxInput;
     let poolInput: PoolStateInput;
     beforeAll(async () => {
@@ -80,7 +80,7 @@ describe('weighted exit test', () => {
         );
     });
 
-    describe('unbalanced exit', async () => {
+    describe('remove liquidity unbalanced', async () => {
         let input: Omit<RemoveLiquidityUnbalancedInput, 'amountsOut'>;
         let amountsOut: InputAmount[];
         beforeAll(() => {
@@ -95,16 +95,16 @@ describe('weighted exit test', () => {
                 kind: RemoveLiquidityKind.Unbalanced,
             };
         });
-        test('removing liquidity with wrapped', async () => {
+        test('with wrapped', async () => {
             const removeLiquidityInput = {
                 ...input,
                 amountsOut: amountsOut.slice(0, 1),
             };
-            const removeLiquidityOutput = await doExit({
+            const removeLiquidityOutput = await doRemoveLiquidity({
                 ...txInput,
                 removeLiquidityInput,
             });
-            assertUnbalancedExit(
+            assertRemoveLiquidityUnbalanced(
                 txInput.client.chain?.id as number,
                 txInput.poolStateInput,
                 removeLiquidityInput,
@@ -112,17 +112,17 @@ describe('weighted exit test', () => {
                 txInput.slippage,
             );
         });
-        test('removing liquidity with native', async () => {
+        test('with native', async () => {
             const removeLiquidityInput = {
                 ...input,
                 amountsOut: amountsOut.slice(0, 1),
                 exitWithNativeAsset: true,
             };
-            const removeLiquidityOutput = await doExit({
+            const removeLiquidityOutput = await doRemoveLiquidity({
                 ...txInput,
                 removeLiquidityInput,
             });
-            assertUnbalancedExit(
+            assertRemoveLiquidityUnbalanced(
                 txInput.client.chain?.id as number,
                 txInput.poolStateInput,
                 removeLiquidityInput,
@@ -132,7 +132,7 @@ describe('weighted exit test', () => {
         });
     });
 
-    describe('single asset exit', () => {
+    describe('remove liquidity single asset', () => {
         let input: RemoveLiquiditySingleTokenInput;
         beforeAll(() => {
             const bptIn: InputAmount = {
@@ -149,13 +149,13 @@ describe('weighted exit test', () => {
                 kind: RemoveLiquidityKind.SingleAsset,
             };
         });
-        test('removing liquidity with wrapped', async () => {
-            const removeLiquidityOutput = await doExit({
+        test('with wrapped', async () => {
+            const removeLiquidityOutput = await doRemoveLiquidity({
                 ...txInput,
                 removeLiquidityInput: input,
             });
 
-            assertSingleTokenExit(
+            assertRemoveLiquiditySingleToken(
                 txInput.client.chain?.id as number,
                 txInput.poolStateInput,
                 input,
@@ -164,17 +164,17 @@ describe('weighted exit test', () => {
             );
         });
 
-        test('removing liquidity with native', async () => {
+        test('with native', async () => {
             const removeLiquidityInput = {
                 ...input,
                 exitWithNativeAsset: true,
             };
-            const removeLiquidityOutput = await doExit({
+            const removeLiquidityOutput = await doRemoveLiquidity({
                 ...txInput,
                 removeLiquidityInput,
             });
 
-            assertSingleTokenExit(
+            assertRemoveLiquiditySingleToken(
                 txInput.client.chain?.id as number,
                 txInput.poolStateInput,
                 removeLiquidityInput,
@@ -184,7 +184,7 @@ describe('weighted exit test', () => {
         });
     });
 
-    describe('proportional exit', () => {
+    describe('remove liquidity proportional', () => {
         let input: RemoveLiquidityProportionalInput;
         beforeAll(() => {
             const bptIn: InputAmount = {
@@ -200,12 +200,12 @@ describe('weighted exit test', () => {
             };
         });
         test('with tokens', async () => {
-            const removeLiquidityOutput = await doExit({
+            const removeLiquidityOutput = await doRemoveLiquidity({
                 ...txInput,
                 removeLiquidityInput: input,
             });
 
-            assertProportionalExit(
+            assertRemoveLiquidityProportional(
                 txInput.client.chain?.id as number,
                 txInput.poolStateInput,
                 input,
@@ -218,11 +218,11 @@ describe('weighted exit test', () => {
                 ...input,
                 useNativeAssetAsWrappedAmountIn: true,
             };
-            const removeLiquidityOutput = await doExit({
+            const removeLiquidityOutput = await doRemoveLiquidity({
                 ...txInput,
                 removeLiquidityInput,
             });
-            assertProportionalExit(
+            assertRemoveLiquidityProportional(
                 txInput.client.chain?.id as number,
                 txInput.poolStateInput,
                 removeLiquidityInput,
